+++ kusto/conn.go
@@ -79,7 +79,7 @@ type connOptions struct {
 
 // query makes a query for the purpose of extracting data from Kusto. Context can be used to set
 // a timeout or cancel the query. Queries cannot take longer than 5 minutes.
-func (c *Conn) query(ctx context.Context, db string, query Stmt, options *queryOptions) (execResp, error) {
+func (c *Conn) query(ctx context.Context, db string, query Statement, options *queryOptions) (execResp, error) {
 	if strings.HasPrefix(strings.TrimSpace(query.String()), ".") {
 		return execResp{}, errors.ES(errors.OpQuery, errors.KClientArgs, "a Stmt to Query() cannot begin with a period(.), only Mgmt() calls can do that").SetNoRetry()
 	}
@@ -92,7 +92,7 @@ func (c *Conn) mgmt(ctx context.Context, db string, query Stmt, options *mgmtOpt
 	return c.execute(ctx, execMgmt, db, query, *options.requestProperties)
 }
 
-func (c *Conn) queryToJson(ctx context.Context, db string, query Stmt, options *queryOptions) (string, error) {
+func (c *Conn) queryToJson(ctx context.Context, db string, query Statement, options *queryOptions) (string, error) {
 	_, _, _, body, e := c.doRequest(ctx, execQuery, db, query, *options.requestProperties)
 	if e != nil {
 		return "", e
@@ -114,7 +114,7 @@ type execResp struct {
 	frameCh    chan frames.Frame
 }
 
-func (c *Conn) execute(ctx context.Context, execType int, db string, query Stmt, properties requestProperties) (execResp, error) {
+func (c *Conn) execute(ctx context.Context, execType int, db string, query Statement, properties requestProperties) (execResp, error) {
 	op, reqHeader, respHeader, body, e := c.doRequest(ctx, execType, db, query, properties)
 	if e != nil {
 		return execResp{}, e
@@ -135,7 +135,7 @@ func (c *Conn) execute(ctx context.Context, execType int, db string, query Stmt,
 	return execResp{reqHeader: reqHeader, respHeader: respHeader, frameCh: frameCh}, nil
 }
 
-func (c *Conn) doRequest(ctx context.Context, execType int, db string, query Stmt, properties requestProperties) (errors.Op, http.Header, http.Header,
+func (c *Conn) doRequest(ctx context.Context, execType int, db string, query Statement, properties requestProperties) (errors.Op, http.Header, http.Header,
 	io.ReadCloser, error) {
 	err := c.validateEndpoint()
 	var op errors.Op
@@ -153,10 +153,17 @@ func (c *Conn) doRequest(ctx context.Context, execType int, db string, query Stm
 
 	switch execType {
 	case execQuery, execMgmt:
+		var err error
+		var csl string
+		if query.SupportsInlineParameters() || properties.QueryParameters.Count() == 0 {
+			csl = query.String()
+		} else {
+			csl = fmt.Sprintf("%s\n%s", properties.QueryParameters.ToDeclarationString(), query.String())
+		}
 		err = json.NewEncoder(buff).Encode(
 			queryMsg{
 				DB:         db,
-				CSL:        query.String(),
+				CSL:        csl,
 				Properties: properties,
 			},
 		)