+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [Unreleased]
+### Added
+* Added kql.Builder struct for safe building of KQL statements from variables without use of 'Unsafe' mode.
+* Simpler handling of query parameters using kql.Parameters struct. 
 
 ## [0.11.2] - 2023-03-14
 ### Fixed